requirements.txt
opencv-python >= 4.5.1.48
depthai >= 2.15.4
shapely

pyzmq
panda3d